API_URL + Svelte static adapter + Remove https from dotnet server
More information about dotnet server can be found at:
https://docs.microsoft.com/en-us/aspnet/core/host-and-deploy/linux-nginx?view=aspnetcore-6.0

There are still changes that need to be made. Security has not
been a priority when making these changes. I just want it to work
then I can worry about (the probably more important) other stuff.

I spent a long time wondering why nginx could not see the local
webserver hosted on port 5147. I got the error message bad gateway
502 and thought it was because dotnet refused to answer the proxy.
However it is because of SELinux. More information about this can
be found at:
https://www.redhat.com/sysadmin/setting-reverse-proxies-nginx

But basically SELinux only allows nginx access to port 80 and 443,
however as we want it to access more ports, we have to make some
changes. That can be done with:
-> sudo setsebool -P httpd_can_network_connect 1